+++ CHANGELOG.md
@@ -58,6 +58,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     * Fix styles of `export excel` and `cols config` buttons.
 * `dynamic-actions` mixin:
     * Fix `classNames` property computing.
+* `flexberry-file` component:
+    * Fix preview click action.
 
 ### Removed
 * The following templates: `sitemap`, `sitemap-node` (and mobile), `sitemap-node-content`.